+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 ## ✨ Features
 
 * show/hide InApp Browser on cozy-intent call ([PR #179](https://github.com/cozy/cozy-pass-mobile/pull/179))
+* Intercept redirection/reloading and refresh only WebView ([PR #187](https://github.com/cozy/cozy-react-native/pull/187))
 
 ## 🐛 Bug Fixes
 
@@ -24,8 +25,8 @@
 * Fix possible cases of failed handshake preventing webview apps to communicate with the native code ([PR #168](https://github.com/cozy/cozy-pass-mobile/pull/168))
 
 ## 🔧 Tech
-* Improve webview env injection ([PR #167](https://github.com/cozy/cozy-pass-mobile/pull/167))
 
+* Improve webview env injection ([PR #167](https://github.com/cozy/cozy-pass-mobile/pull/167))
 * Run iOS installation on CI
 
 # 0.0.7